.update vscode generator

+++ pym/bob/generators/VsCode.py
@@ -62,70 +62,57 @@ def generateFile(entries, fileName):
         fileName.write(e + "\n")
     fileName.close()
 
-def addRunStep(outFile, num, cmd, directory):
-   outFile.write('<valuemap type="QVariantMap" key="ProjectExplorer.Target.RunConfiguration.' + str(num) + '">')
-   outFile.write(' <value type="QString" key="ProjectExplorer.CustomExecutableRunConfiguration.Arguments"></value>')
-   outFile.write(' <value type="QString" key="ProjectExplorer.CustomExecutableRunConfiguration.Executable">' + cmd + '</value>')
-   outFile.write(' <value type="QString" key="ProjectExplorer.CustomExecutableRunConfiguration.WorkingDirectory">' + directory + '</value>')
-   outFile.write(' <value type="QString" key="ProjectExplorer.ProjectConfiguration.DefaultDisplayName">' + cmd + '</value>')
-   outFile.write(' <value type="QString" key="ProjectExplorer.ProjectConfiguration.DisplayName"></value>')
-   outFile.write(' <value type="QString" key="ProjectExplorer.ProjectConfiguration.Id">ProjectExplorer.CustomExecutableRunConfiguration</value>')
-   outFile.write(' <value type="bool" key="RunConfiguration.UseCppDebugger">false</value>')
-   outFile.write(' <value type="bool" key="RunConfiguration.UseCppDebuggerAuto">true</value>')
-   outFile.write(' <value type="bool" key="RunConfiguration.UseMultiProcess">false</value>')
-   outFile.write(' <value type="bool" key="RunConfiguration.UseQmlDebugger">false</value>')
-   outFile.write(' <value type="bool" key="RunConfiguration.UseQmlDebuggerAuto">true</value>')
-   outFile.write('</valuemap>')
+def addRunSteps(outFile, runTargets):
+    if runTargets:
+        print("\nNot implemented: {} add run-targets {}".format(outFile.name, runTargets))
 
 RunStep = namedtuple("RunStep", "dir cmd")
 
-def addRunSteps(outFile, runTargets):
-    runTargetCnt = 0
-    for directory,cmd in runTargets:
-        addRunStep(outFile, runTargetCnt, cmd, directory)
-        runTargetCnt += 1
-    outFile.write('<value type="int" key="ProjectExplorer.Target.RunConfigurationCount">' + str(runTargetCnt) + '</value>')
-
 def addBuildConfig(outFile, num, name, buildArgs, buildMeFile):
-    outFile.write('<valuemap type="QVariantMap" key="ProjectExplorer.Target.BuildConfiguration.' + str(num) + '">\n')
-    outFile.write('<value type="QString" key="ProjectExplorer.BuildConfiguration.BuildDirectory">' + cwd() + '</value>\n')
-    outFile.write('<valuemap type="QVariantMap" key="ProjectExplorer.BuildConfiguration.BuildStepList.0">\n')
-    outFile.write(' <valuemap type="QVariantMap" key="ProjectExplorer.BuildStepList.Step.0">\n')
-    outFile.write('  <value type="bool" key="ProjectExplorer.BuildStep.Enabled">true</value>\n')
+    outFile.write("    {\n")
+    outFile.write("      \"label\": \"{}\",\n".format(name))
+    outFile.write("      \"type\": \"process\",\n")
     if isWindows():
-        outFile.write('  <value type="QString" key="ProjectExplorer.ProcessStep.Arguments">-msys2 -defterm -no-start -use-full-path -here -c "PATH=' + os.getenv('PATH') + ' sh ' + buildMeFile + ' -v ' + buildArgs + '"</value>\n')
         if os.getenv('WD') is None:
             raise BuildError("Cannot create QtCreator project for Windows! MSYS2 must be started by msys2_shell.cmd script!")
-        outFile.write('  <value type="QString" key="ProjectExplorer.ProcessStep.Command">' + os.path.normpath(os.path.join(os.getenv('WD').replace('\\', '/'), '..', '..', 'msys2_shell.cmd')) + '</value>\n')
+        outFile.write("      \"command\": \"{}\",\n".format(os.path.normpath(os.path.join(os.getenv('WD').replace('\\', '/'), "sh.exe"))))
     else:
-        outFile.write('  <value type="QString" key="ProjectExplorer.ProcessStep.Arguments">' + buildArgs + '</value>\n')
-        outFile.write('  <value type="QString" key="ProjectExplorer.ProcessStep.Command">' + buildMeFile + '</value>\n')
-    outFile.write('  <value type="QString" key="ProjectExplorer.ProcessStep.WorkingDirectory">%{buildDir}</value>\n')
-    outFile.write('  <value type="QString" key="ProjectExplorer.ProjectConfiguration.DefaultDisplayName">' + name + '</value>\n')
-    outFile.write('  <value type="QString" key="ProjectExplorer.ProjectConfiguration.DisplayName"></value>\n')
-    outFile.write('  <value type="QString" key="ProjectExplorer.ProjectConfiguration.Id">ProjectExplorer.ProcessStep</value>\n')
-    outFile.write(' </valuemap>\n')
-    outFile.write(' <value type="int" key="ProjectExplorer.BuildStepList.StepsCount">1</value>\n')
-    outFile.write(' <value type="QString" key="ProjectExplorer.ProjectConfiguration.DefaultDisplayName">Build</value>\n')
-    outFile.write(' <value type="QString" key="ProjectExplorer.ProjectConfiguration.DisplayName"></value>\n')
-    outFile.write(' <value type="QString" key="ProjectExplorer.ProjectConfiguration.Id">ProjectExplorer.BuildSteps.Build</value>\n')
-    outFile.write('</valuemap>\n')
-    outFile.write('<value type="int" key="ProjectExplorer.BuildConfiguration.BuildStepListCount">1</value>\n')
-    outFile.write('<value type="bool" key="ProjectExplorer.BuildConfiguration.ClearSystemEnvironment">false</value>\n')
-    outFile.write('<valuelist type="QVariantList" key="ProjectExplorer.BuildConfiguration.UserEnvironmentChanges">\n')
-    outFile.write(' <value type="QString">' + "PATH=" + os.environ['PATH'] + '</value>\n')
-    outFile.write('</valuelist>\n')
-    outFile.write('<value type="QString" key="ProjectExplorer.ProjectConfiguration.DefaultDisplayName">Vorgabe</value>\n')
-    outFile.write('<value type="QString" key="ProjectExplorer.ProjectConfiguration.DisplayName">' + name + '</value>\n')
-    outFile.write('<value type="QString" key="ProjectExplorer.ProjectConfiguration.Id">GenericProjectManager.GenericBuildConfiguration</value>\n')
-    outFile.write('<valuelist type="QVariantList" key="UserStickyKeys">\n')
-    outFile.write(' <value type="QString">ProjectExplorer.BuildConfiguration.UserEnvironmentChanges</value>\n')
-    outFile.write(' <value type="QString">UserStickyKeys</value>\n')
-    outFile.write('</valuelist>\n')
-    outFile.write('</valuemap>\n')
+        outFile.write("      \"command\": \"sh\",\n")
+    outFile.write("      \"args\": [\n")
+    outFile.write("        \"{}\",\n".format(buildMeFile))
+    outFile.write("        \"-v\",\n")
+    for i in buildArgs.split(' '):
+        outFile.write("        \"{}\",\n".format(i))
+    outFile.write("      ],\n")
+    outFile.write("      \"group\": {\n")
+    outFile.write("        \"kind\": \"build\",\n")
+    outFile.write("        \"isDefault\": true\n")
+    outFile.write("      }\n")
+    outFile.write("    },\n")
 
 def addBuildSteps(outFile, buildMeFile, buildConfigs):
-    addBuildConfig(outFile, 0, "Bob dev ", "" , buildMeFile)
+    outFile.write("{\n")
+    outFile.write("  \"version\": \"2.0.0\",\n")
+    outFile.write("  \"options\":{\n")
+    outFile.write("    \"shell\":{\n")
+    outFile.write("      \"executable\": \"{}\",\n".format(os.path.normpath(os.path.join(os.getenv('WD').replace('\\', '/'), '..', '..', 'msys2_shell.cmd'))))
+    outFile.write("      \"args\": [\n")
+    outFile.write("        \"-msys2\",\n")
+    outFile.write("        \"-defterm\",\n")
+    outFile.write("        \"-no-start\",\n")
+    outFile.write("        \"-use-full-path\",\n")
+    outFile.write("        \"-here\",\n")
+    outFile.write("        \"-c\",\n")
+    outFile.write("      ]\n")
+    outFile.write("    },\n")
+    outFile.write("    \"env\": {\n")
+    outFile.write("      \"PATH\": \"{}\",\n".format(os.getenv('PATH')))
+    outFile.write("    },\n")
+    outFile.write("    \"cwd\": \"{}\",\n".format(cwd()))
+    outFile.write("  },\n")
+    outFile.write("  \"tasks\": [\n")
+
+    addBuildConfig(outFile, 0, "Bob dev", "" , buildMeFile)
     addBuildConfig(outFile, 1, "Bob dev (force)", "-f", buildMeFile)
     addBuildConfig(outFile, 2, "Bob dev (no checkout)", "-b", buildMeFile)
     addBuildConfig(outFile, 3, "Bob dev (no deps)", "-n", buildMeFile)
@@ -136,13 +123,9 @@ def addBuildSteps(outFile, buildMeFile, buildConfigs):
     for name,flags in buildConfigs:
         addBuildConfig(outFile, count, name, flags, buildMeFile)
         count += 1
-
-    outFile.write('<value type="int" key="ProjectExplorer.Target.BuildConfigurationCount">' + str(count) + '</value>\n')
-    outFile.write('<valuelist type="QVariantList" key="UserStickyKeys">\n')
-    outFile.write(' <value type="QString">ProjectExplorer.BuildConfiguration.BuildStepListCount</value>\n')
-    outFile.write(' <value type="QString">ProjectExplorer.BuildConfiguration.UserEnvironmentChanges</value>\n')
-    outFile.write(' <value type="QString">ProjectExplorer.ProjectConfiguration.DisplayName</value>\n')
-    outFile.write('</valuelist>\n')
+    
+    outFile.write("  ]\n")
+    outFile.write("}\n")
 
 def compareAndRenameFileIfNotEqual(orig, new):
     if os.path.exists(orig):
@@ -233,6 +216,8 @@ def vscodeProjectGenerator(package, argv, extra, bobRoot):
         removePath(destination)
     if not os.path.exists(destination):
         os.makedirs(destination)
+    if not os.path.exists(os.path.join(destination, ".vscode")):
+        os.makedirs(os.path.join(destination, ".vscode"))
 
     # Path structure:
     # destination
@@ -243,9 +228,10 @@ def vscodeProjectGenerator(package, argv, extra, bobRoot):
     # create a 'flat' source tree
     symlinkDir = os.path.join(destination, projectName)
     if not args.update:
-       if os.path.exists(symlinkDir):
-          shutil.rmtree(symlinkDir)
-       os.makedirs(symlinkDir)
+        if not isWindows():
+            if os.path.exists(symlinkDir):
+                shutil.rmtree(symlinkDir)
+            os.makedirs(symlinkDir)
 
     # lists for storing all found sources files / include directories / defines
     sList = []
@@ -255,6 +241,10 @@ def vscodeProjectGenerator(package, argv, extra, bobRoot):
 
     sList.append("{")
     sList.append("  \"folders\": [")
+    sList.append("    {")
+    sList.append("      \"name\": \".\",")
+    sList.append("      \"path\": \".\"")
+    sList.append("    },")
 
     # now go through all checkout dirs to find all header / include files
     for name,path in OrderedDict(sorted(dirs, key=lambda t: t[1])).items():
@@ -333,7 +323,7 @@ def vscodeProjectGenerator(package, argv, extra, bobRoot):
         buildMe = []
         buildMe.append("#!/bin/sh")
         buildMe.append('bob dev "$@" ' + extra + ' ' + quote(project))
-        projectCmd = "bob project -n " + extra + " qt-creator " + quote(project) + \
+        projectCmd = "bob project -n " + extra + " vscode " + quote(project) + \
             " -u --destination " + quote(destination) + ' --name ' + quote(projectName)
         # only add arguments which are relevant for .files or .includes. All other files are only modified if not build with
         # update only.
@@ -356,77 +346,6 @@ def vscodeProjectGenerator(package, argv, extra, bobRoot):
         generateFile(buildMe, buildMeFile)
         os.chmod(buildMeFile, stat.S_IRWXU | stat.S_IRGRP | stat.S_IWGRP |
             stat.S_IROTH | stat.S_IWOTH)
-        # Generate the creator.shared file using a template and modify some settings
-        sharedFile = os.path.join(destination, projectName  + ".creator.shared")
-
-        template_head = """<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE QtCreatorProject>
-<qtcreator>
- <data>
-  <variable>ProjectExplorer.Project.ActiveTarget</variable>
-  <value type="int">0</value>
- </data>
- <data>
-  <variable>ProjectExplorer.Project.EditorSettings</variable>
-  <valuemap type="QVariantMap">
-   <value type="bool" key="EditorConfiguration.AutoIndent">true</value>
-   <value type="bool" key="EditorConfiguration.AutoSpacesForTabs">false</value>
-   <value type="bool" key="EditorConfiguration.CamelCaseNavigation">true</value>
-   <valuemap type="QVariantMap" key="EditorConfiguration.CodeStyle.0">
-    <value type="QString" key="language">Cpp</value>
-    <valuemap type="QVariantMap" key="value">
-     <value type="QByteArray" key="CurrentPreferences">CppGlobal</value>
-    </valuemap>
-   </valuemap>
-   <valuemap type="QVariantMap" key="EditorConfiguration.CodeStyle.1">
-    <value type="QString" key="language">QmlJS</value>
-    <valuemap type="QVariantMap" key="value">
-     <value type="QByteArray" key="CurrentPreferences">QmlJSGlobal</value>
-    </valuemap>
-   </valuemap>
-   <value type="int" key="EditorConfiguration.CodeStyle.Count">2</value>
-   <value type="QByteArray" key="EditorConfiguration.Codec">UTF-8</value>
-   <value type="bool" key="EditorConfiguration.ConstrainTooltips">false</value>
-   <value type="int" key="EditorConfiguration.IndentSize">4</value>
-   <value type="bool" key="EditorConfiguration.KeyboardTooltips">false</value>
-   <value type="int" key="EditorConfiguration.MarginColumn">80</value>
-   <value type="bool" key="EditorConfiguration.MouseHiding">true</value>
-   <value type="bool" key="EditorConfiguration.MouseNavigation">true</value>
-   <value type="int" key="EditorConfiguration.PaddingMode">1</value>
-   <value type="bool" key="EditorConfiguration.ScrollWheelZooming">true</value>
-   <value type="bool" key="EditorConfiguration.ShowMargin">false</value>
-   <value type="int" key="EditorConfiguration.SmartBackspaceBehavior">0</value>
-   <value type="bool" key="EditorConfiguration.SmartSelectionChanging">true</value>
-   <value type="bool" key="EditorConfiguration.SpacesForTabs">true</value>
-   <value type="int" key="EditorConfiguration.TabKeyBehavior">0</value>
-   <value type="int" key="EditorConfiguration.TabSize">4</value>
-   <value type="bool" key="EditorConfiguration.UseGlobal">true</value>
-   <value type="int" key="EditorConfiguration.Utf8BomBehavior">1</value>
-   <value type="bool" key="EditorConfiguration.addFinalNewLine">true</value>
-   <value type="bool" key="EditorConfiguration.cleanIndentation">true</value>
-   <value type="bool" key="EditorConfiguration.cleanWhitespace">true</value>
-   <value type="bool" key="EditorConfiguration.inEntireDocument">false</value>
-  </valuemap>
- </data>
- <data>
-  <variable>ProjectExplorer.Project.PluginSettings</variable>
-  <valuemap type="QVariantMap"/>
- </data>"""
-
-        template_foot = """
- <data>
-  <variable>ProjectExplorer.Project.TargetCount</variable>
-  <value type="int">1</value>
- </data>
- <data>
-  <variable>ProjectExplorer.Project.Updater.FileVersion</variable>
-  <value type="int">18</value>
- </data>
- <data>
-  <variable>Version</variable>
-  <value type="int">18</value>
- </data>
-</qtcreator>"""
 
         # find all executables in package dir
         runTargets = []
@@ -437,25 +356,18 @@ def vscodeProjectGenerator(package, argv, extra, bobRoot):
                 for filename in filenames:
                     try:
                         ftype = magic.from_file(os.path.join(root, filename))
-                        if 'executable' in ftype and 'x86' in ftype:
+                        if 'executable' in ftype and 'x86' in ftype and (isWindows() and "exe" in filename or not isWindows()):
                             runTargets.append(RunStep(os.path.join(cwd(), root), filename))
                     except OSError:
                         pass
 
+        # Generate the tasks.json file using a template and modify some settings
+        sharedFile = os.path.join(os.path.join(destination, ".vscode"), "tasks.json")
         with open(sharedFile, 'w') as sharedFile:
-            sharedFile.write(template_head)
-            sharedFile.write(' <data>\n')
-            sharedFile.write('  <variable>ProjectExplorer.Project.Target.0</variable>\n')
-            sharedFile.write('  <valuemap type="QVariantMap">\n')
-            sharedFile.write('   <value type="int" key="ProjectExplorer.Target.ActiveBuildConfiguration">0</value>\n')
-            sharedFile.write('   <value type="int" key="ProjectExplorer.Target.ActiveDeployConfiguration">0</value>\n')
-            sharedFile.write('   <value type="int" key="ProjectExplorer.Target.ActiveRunConfiguration">0</value>\n')
             addBuildSteps(sharedFile, buildMeFile, args.buildCfg)
-            sharedFile.write('   <value type="int" key="ProjectExplorer.Target.DeployConfigurationCount">0</value>\n')
-            sharedFile.write('   <valuemap type="QVariantMap" key="ProjectExplorer.Target.PluginSettings"/>\n')
-            addRunSteps(sharedFile, runTargets)
-            sharedFile.write('  </valuemap>\n')
-            sharedFile.write(' </data>\n')
-            sharedFile.write(template_foot)
 
+        # Generate the launch.json file using a template and modify some settings
+        sharedFile = os.path.join(os.path.join(destination, ".vscode"), "launch.json")
+        with open(sharedFile, 'w') as sharedFile:
+            addRunSteps(sharedFile, runTargets)
